Fix regressions - isolate tests
